requirements.txt
openpyxl==3.1.2
pandas==2.1.1
pdfplumber==0.10.2
python-dotenv==1.0.0